Fixed Buttons & Spill
Fixed the "save" and "exit" buttons on the settings screen so that now they work!
Added the sprite for the spills and corresponding animations.
Added a profile/reaction to when there is a spill, and also added other reactions such as content and sad (for like when the day ends)